requirements.txt
APScheduler==3.9.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.0.2
gevent==22.10.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-gitlab==3.14.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-toolbelt==1.0.0
six==1.16.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.13
Werkzeug==2.2.3
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
